Remove unnecessary data mapping.
The mapping that converted DB column names into camelCase names for output is
now done in SQL.

Also only one select statement is used now, as it's no great inefficiency to
always send the block ID along with the data even when it's not going to be
shown.

Fixed a silly double-request bug when toggling 'Block only'.