Load all chars at startup, and filter in JS.
The biggest bottleneck when searching via SQL was transferring the result set
from the database to the application – although queries might only take a few
milliseconds to actually perform, the transfer might take upwards of 100ms.

To work around this, the entire data set is loaded into memory upon application
start, and searches are performed on it in pure JavaScript. The data set is not
particularly large, and JS is fast enough, that overall an approximately 10×
gain in performance is achieved.