Use enable click-to-copy; change to /ucdb as root
This commit also changes how clicks are handled on characters in the
results list. Previously it would just select them; now it just copies
them directly to the clipboard. To enable this, the app needed to be
running under HTTPS, hence the second part of this update.

Since I have no certificate for https://ucdb.andyf.me/, the app is now
configured to run under https://andyf.me/ucdb instead, which meant
rewriting the paths for assets and searching.